+++ gateways/stripe/script.js
@@ -10,26 +10,24 @@
                 return;
             }
 
-            /***********************************************************/
-            /* Configure Stripe
-            /***********************************************************/
+            var amount = parseInt(ShoppingCart.calculateTotalPriceIncludingTaxesAndShipping().toString().replace('.', ''));
+
             var stripeHandler = StripeCheckout.configure({
                 key: ShoppingCart.settings.payment.methods.stripe.publicKey,
                 token: function(token, args) {
                     var order = {
                         products: storejs.get('grav-shoppingcart-basket-data'),
-                        address: storejs.get('grav-shoppingcart-person-address'),
+                        data: storejs.get('grav-shoppingcart-checkout-form-data'),
                         shipping: storejs.get('grav-shoppingcart-shipping-method'),
-                        payment: storejs.get('grav-shoppingcart-payment-method'),
+                        payment: 'stripe',
                         token: storejs.get('grav-shoppingcart-order-token').token,
                         extra: { 'stripeToken': token.id },
-                        amount: ShoppingCart.totalOrderPrice.toString(),
+                        amount: amount,
                         gateway: ShoppingCart.gateway
                     };
 
-
                     jQuery.ajax({
-                        url: ShoppingCart.settings.baseURL + ShoppingCart.settings.urls.saveOrderURL + '?task=pay',
+                        url: ShoppingCart.settings.baseURL + ShoppingCart.settings.urls.save_order_url + '/task:pay',
                         data: order,
                         type: 'POST'
                     })
@@ -46,8 +44,8 @@
             stripeHandler.open({
                 name: ShoppingCart.settings.payment.methods.stripe.name,
                 description: ShoppingCart.settings.payment.methods.stripe.description,
-                email: storejs.get('grav-shoppingcart-person-address').email,
-                amount: ShoppingCart.calculateTotalPriceIncludingTaxesAndShipping().toString().replace('.', ''),
+                email: storejs.get('grav-shoppingcart-checkout-form-data').email,
+                amount: amount,
                 currency: ShoppingCart.settings.general.currency
             });
         });