+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# v1.1.0
+## 05-12-2016
+
+1. [](#new)
+    * 1.1 released
+
+# v1.1.0-beta.3
+## 09-08-2016
+
+1. [](#bugfix)
+    * Fix issue with Stripe amount format
+    * Fix issue in `enabled` state conflicting with main plugin
+
+# v1.1.0-beta.2
+## 30-04-2016
+
+1. [](#new)
+    * Use the new URL callback structure for ShoppingCart 1.1
+    * Namespace
+    * Update dependencies
+
 # v1.0.5
 ## 28-04-2016
 
@@ -15,6 +36,12 @@
 
 1. [](#bugfix)
     * Fix the dependencies format
+
+# v1.0.2
+## 20-03-2016
+
+1. [](#bugfix)
+    * Fix issue with require
 
 # v1.0.1
 ## 01-03-2016